updated script

+++ android-lan.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 [[ -n $2 ]] && echo 'Пробелы запрещены!' && exit 1
-[[ -f ~/switch-lan-play/src/build/lanplay ]] && sudo ~/switch-lan-play/src/build/lanplay --relay-server-addr $1 && exit
+if [[ -f ~/switch-lan-play/src/build/lanplay ]]; then
+    [[ -z $1 ]] && echo -n 'Введите адрес сервера:' && read server || server="$1"
+    [[ -z $server ]] && echo 'Адрес сервера не может быть пустым!' && exit 1
+    sudo ~/switch-lan-play/src/build/lanplay --relay-server-addr $server && exit
 apt update
 apt upgrade -y
 apt install libpcap -y
